components/server/requirements.txt
llama-index==0.6.0
Pillow==9.5.0
PyPDF2==3.0.1
nltk==3.8.1
google-cloud-firestore==2.11.1
fastapi==0.95.1
uvicorn==0.22.0
gunicorn==20.1.0
